fsimagestorage initiated
